Implement Wave System
Created a simple wave system for playtest.

+++ Proyekt-Game/Proyekt/Assets/INEAIP/Runtime/Package/Health/IrminBaseHealthSystem.cs
@@ -76,7 +76,7 @@ namespace irminNavmeshEnemyAiUnityPackage
             }
         }
 
-        public void ReAwaken(int pMaxHealth)
+        public void ReAwaken(float pMaxHealth)
         {
             _maxHealth = pMaxHealth;
             Awake();
